Python/Gitlab/requirements.txt
python-gitlab
dumper
requests
